% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_URF.R
\docType{data}
\name{expr_URF}
\alias{expr_URF}
\title{Time-restricted feeding expression data of unrestricted feeding (URF) group}
\format{
A data frame with 13169 rows and 62 variables. Each row is the gene expression value. Each column represents one sample.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE129843}
}
\usage{
expr_URF
}
\description{
Short-Term Time-Restricted Feeding Alters Rhythmicity of Lipid and Amino Acid Metabolites in Overweight Men expression data of unrestricted feeding (URF) group
}
\references{
Lundell LS, Parr EB, Devlin BL, Ingerslev LR et al. Time-restricted feeding alters lipid and amino acid metabolite rhythmicity without perturbing clock gene expression. Nat Commun 2020 Sep 16;11(1):4643. PMID: 32938935
}
\keyword{datasets}
