% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_TRF.R
\docType{data}
\name{pheno_TRF}
\alias{pheno_TRF}
\title{Subjects information for time-restricted feeding (TRF) group of the time-restricted feeding data.}
\format{
A data frame with 63 rows and 3 columns. Each row represents one observation. First column is the sample type. Second column is the subject ID. Third column is the time in hours.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE129843}
}
\usage{
pheno_TRF
}
\description{
Subjects information for Short-Term Time-Restricted Feeding Alters Rhythmicity of Lipid and Amino Acid Metabolites in Overweight Men data in time-restricted feeding (TRF) group.
}
\references{
Lundell LS, Parr EB, Devlin BL, Ingerslev LR et al. Time-restricted feeding alters lipid and amino acid metabolite rhythmicity without perturbing clock gene expression. Nat Commun 2020 Sep 16;11(1):4643. PMID: 32938935
}
\keyword{datasets}
