% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_rhythmicity_with_prior.R
\name{bayes_rhythmicity_prior}
\alias{bayes_rhythmicity_prior}
\title{Bayesian Method for Detecting Circadian Rhythmicity with Prior Information.}
\usage{
bayes_rhythmicity_prior(
  yy,
  tt,
  period = 24,
  M = 3000,
  alpha00_nc = 0,
  alpha00_c = 0,
  beta10_c = 0,
  beta20_c = 0,
  tau2a_nc = 10,
  tau2a_c = 10,
  tau2b1_c = 10,
  tau2b2_c = 10,
  mu0_nc = 0,
  mu0_c = 0,
  nu0_nc = 0,
  nu0_c = 0,
  a_p = 1,
  b_p = 1,
  prior_gene = rep(0, nrow(yy)),
  prior_FDR = 0.05
)
}
\arguments{
\item{yy}{Gene expression value data matrix. Rows are genes' expression values, columns are observation time points.}

\item{tt}{Time vector.}

\item{period}{Period of your data. Default is 24.}

\item{M}{number of MCMC iterations. Default is 3000.}

\item{alpha00_nc}{hyper parameter for non-circadian model. Default is 0.}

\item{alpha00_c}{hyper parameter for circadian model. Default is 0.}

\item{beta10_c}{hyper parameter for circadian model. Default is 0.}

\item{beta20_c}{hyper parameter for circadian model. Default is 0.}

\item{tau2a_nc}{hyper parameter for non-circadian model. Default is 10.}

\item{tau2a_c}{hyper parameter for circadian model. Default is 10.}

\item{tau2b1_c}{hyper parameter for circadian model. Default is 10.}

\item{tau2b2_c}{hyper parameter for circadian model. Default is 10.}

\item{mu0_nc}{hyper parameter for non-circadian model. Default is 0.}

\item{mu0_c}{hyper parameter for circadian model. Default is 0.}

\item{nu0_nc}{hyper parameter for non-circadian model. Default is 0.}

\item{nu0_c}{hyper parameter for circadian model. Default is 0.}

\item{a_p}{hyper parameter. Default is 1.}

\item{b_p}{hyper parameter. Default is 1.}

\item{prior_gene}{Optional prior information. A vector of 0 and 1 indicators. 0 indicates non-circadian gene, 1 indicates circadian gene. Default is all non-circadian genes (all zeros).}

\item{prior_FDR}{FDR cutoff for the prior data. Default is 5\%.}
}
\value{
A list of Bayesian q-values and running time in seconds.
\item{qvalue}{A vector of q-values for each gene.}
\item{pvalue}{A vector of p-values for each gene.}
\item{alpha0}{A matrix contains all non-circadian genes alpha values for each iteration.}
\item{sigma20}{A matrix contains all non-circadian genes sigma square values for each iteration.}
\item{alpha1}{A matrix contains all circadian genes alpha values for each iteration.}
\item{beta1}{A matrix contains all circadian genes beta1 values for each iteration.}
\item{beta2}{A matrix contains all circadian genes beta2 values for each iteration.}
\item{sigma21}{A matrix contains all circadian genes sigma square values for each iteration.}
\item{r}{Gene indicator. 0 is non-circadian gene, 1 is circadian gene.}
\item{k}{Sum of circadian genes in each MCMC iteration.}
\item{p}{Posterior probability of number of circadian gene.}
\item{timeUsed}{Running time used in seconds.}
}
\description{
Test the significance of circadian curve fitting using Bayesian method with prior information.
}
\details{
Test the significance of circadian curve fitting using Bayesian method with prior information.
}
\examples{
set.seed(32611)
G=1000 
G1=100 
G0=G-G1
A1=3 
Mesor1=3
sigma21=1 
A0=0 
Mesor0=3 
sigma20=1
n=24
tt <- runif(n,0,24)
# generate circadian gene samples #
phi1 <- runif(G1,0,24)
w <- 2*pi/24
EE1 <- A1*cos(w*phi1)
FF1 <- A1*sin(w*phi1) 
yy1 <- matrix(0,G1,n)
for(i in 1:G1){
  yy1[i,] <- rnorm(n, EE1[i]*cos(w*tt) + FF1[i]*sin(w*tt) + Mesor1, sqrt(sigma21))
}  
# generate non-circadian gene samples #
phi0 <- runif(G0,0,24)
w <- 2*pi/24
EE0 <- A0*cos(w*phi0)
FF0 <- A0*sin(w*phi0) 
yy0 <- matrix(0,G0,n)
for(i in 1:G0){
  yy0[i,] <- rnorm(n, EE0[i]*cos(w*tt) + FF0[i]*sin(w*tt) + Mesor0, sqrt(sigma20))
}
yy <- rbind(yy1,yy0)
# Given prior information
prior_info <- c(rep(1,G1), rep(0,G0))
bayes_rhythmicity_prior(yy,tt,M=1000, prior_gene = prior_info)
}
\author{
Zhiguang Huo, Haocheng Ding, Lingsong Meng
}
